token count return appropriate data type

+++ simple_shell/sh.h
@@ -16,7 +16,7 @@ void cmd_mode(void);
 
 void _tokenise(char *buffer, char *delim);
 
-int _count_tok(char *buffer, char *delim);
+size_t _count_tok(char *buffer, char *delim);
 
 size_t _strlen(const char *_string);
